+++ lib/PuppeteerSharp.Tests/PuppeteerTests/PuppeteerLaunchTests.cs
@@ -142,7 +142,7 @@ namespace PuppeteerSharp.Tests.PuppeteerTests
             {
                 var launcher = new Launcher(TestConstants.LoggerFactory);
                 var options = TestConstants.DefaultBrowserOptions();
-                options.Args = options.Args.Concat(new[] {$"--user-data-dir=\"{userDataDir}\""}).ToArray();
+                options.Args = options.Args.Concat(new[] { $"--user-data-dir=\"{userDataDir}\"" }).ToArray();
 
                 using (var browser = await launcher.LaunchAsync(options))
                 {
@@ -162,7 +162,7 @@ namespace PuppeteerSharp.Tests.PuppeteerTests
             {
                 var launcher = new Launcher(TestConstants.LoggerFactory);
                 var options = TestConstants.DefaultBrowserOptions();
-                options.Args = options.Args.Concat(new[] {$"--user-data-dir=\"{userDataDir}\""}).ToArray();
+                options.Args = options.Args.Concat(new[] { $"--user-data-dir=\"{userDataDir}\"" }).ToArray();
 
                 using (var browser = await launcher.LaunchAsync(options))
                 {
@@ -187,7 +187,7 @@ namespace PuppeteerSharp.Tests.PuppeteerTests
             {
                 var launcher = new Launcher(TestConstants.LoggerFactory);
                 var options = TestConstants.DefaultBrowserOptions();
-                options.Args = options.Args.Concat(new[] {$"--user-data-dir=\"{userDataDir}\""}).ToArray();
+                options.Args = options.Args.Concat(new[] { $"--user-data-dir=\"{userDataDir}\"" }).ToArray();
 
                 using (var browser = await launcher.LaunchAsync(options))
                 {
@@ -203,60 +203,6 @@ namespace PuppeteerSharp.Tests.PuppeteerTests
                     await page2.GoToAsync(TestConstants.EmptyPage);
                     Assert.Equal("doSomethingOnlyOnce=true", await page2.EvaluateExpressionAsync("document.cookie"));
                 }
-            }
-        }
-
-        [Fact]
-        public async Task HeadlessShouldBeAbleToReadCookiesWrittenByHeadful()
-        {
-            using (var userDataDir = new TempDirectory())
-            {
-                var launcher = new Launcher(TestConstants.LoggerFactory);
-                var options = TestConstants.DefaultBrowserOptions();
-                options.Args = options.Args.Concat(new[] {$"--user-data-dir=\"{userDataDir}\""}).ToArray();
-                options.Headless = false;
-
-                using (var browser = await launcher.LaunchAsync(options))
-                {
-                    var page = await browser.NewPageAsync();
-                    await page.GoToAsync(TestConstants.EmptyPage);
-                    await page.EvaluateExpressionAsync(
-                        "document.cookie = 'foo=true; expires=Fri, 31 Dec 9999 23:59:59 GMT'");
-                }
-
-                options.Headless = true;
-                using (var browser2 = await Puppeteer.LaunchAsync(options, TestConstants.LoggerFactory))
-                {
-                    var page2 = await browser2.NewPageAsync();
-                    await page2.GoToAsync(TestConstants.EmptyPage);
-                    Assert.Equal("foo=true", await page2.EvaluateExpressionAsync("document.cookie"));
-                }
-            }
-        }
-
-        [Fact]
-        public async Task OOPIFShouldReportGoogleComFrame()
-        {
-            // https://google.com is isolated by default in Chromium embedder.
-            var headfulOptions = TestConstants.DefaultBrowserOptions();
-            headfulOptions.Headless = false;
-            using (var browser = await Puppeteer.LaunchAsync(headfulOptions))
-            using (var page = await browser.NewPageAsync())
-            {
-                await page.GoToAsync(TestConstants.EmptyPage);
-                await page.SetRequestInterceptionAsync(true);
-                page.Request += async (sender, e) => await e.Request.RespondAsync(
-                    new ResponseData { Body = "{ body: 'YO, GOOGLE.COM'}" });
-                await page.EvaluateFunctionHandleAsync(@"() => {
-                    const frame = document.createElement('iframe');
-                    frame.setAttribute('src', 'https://google.com/');
-                    document.body.appendChild(frame);
-                    return new Promise(x => frame.onload = x);
-                }");
-                await page.WaitForSelectorAsync("iframe[src=\"https://google.com/\"]");
-                var urls = Array.ConvertAll(page.Frames, frame => frame.Url);
-                Array.Sort(urls);
-                Assert.Equal(new[] { TestConstants.EmptyPage, "https://google.com/" }, urls);
             }
         }
 
@@ -389,18 +335,6 @@ namespace PuppeteerSharp.Tests.PuppeteerTests
         public async Task ShouldHaveDefaultUrlWhenLaunchingBrowser()
         {
             using (var browser = await Puppeteer.LaunchAsync(TestConstants.DefaultBrowserOptions(), TestConstants.LoggerFactory))
-            {
-                var pages = (await browser.PagesAsync()).Select(page => page.Url);
-                Assert.Equal(new[] { TestConstants.AboutBlank }, pages);
-            }
-        }
-
-        [Fact]
-        public async Task ShouldHaveDefaultUrlWhenLaunchingBrowserWithHeadlessFalse()
-        {
-            var options = TestConstants.DefaultBrowserOptions();
-            options.Headless = false;
-            using (var browser = await Puppeteer.LaunchAsync(options, TestConstants.LoggerFactory))
             {
                 var pages = (await browser.PagesAsync()).Select(page => page.Url);
                 Assert.Equal(new[] { TestConstants.AboutBlank }, pages);
